migrate numerous rep_interchange functions into common
